pyproject.toml
[project]
name = "nonebot-plugin-helpwithpic"
version = "1.6"
description = "nonebot插件-动态帮助图片制作"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
keywords = ["egg", "bacon", "sausage", "tomatoes", "Lobster Thermidor"]
authors = [
  {email = "cub_staryow@outlook.com"},
  {name = "cubstaryow"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

dependencies = [
  "nonebot2",
  "anyio",
  "nonebot-adapter-onebot",
  "nonebot-plugin-localstore",
  "aiohttp",
  "pil_utils",
  "pillow"
]

[project.urls]
repository = "https://github.com/cubstaryow/nonebot-plugin-HelpWithPic"
